realizing I need abstract controller

+++ pe/runner/topology.py
@@ -3,6 +3,7 @@ import jsonpickle
 from tqdm import tqdm
 from pe.config.parse import TopologyConfig
 from pe.runner.agent import Agent, Node, Proxy
+import time
 
 
 class Topology:
@@ -17,6 +18,7 @@ class Topology:
         for node in self.config.nodes:
             self.nodes.append(Node(node, is_local))
         self.proxy = Proxy(self.config.proxy, is_local)
+        self.is_local = is_local
     
     @property
     def agents(self) -> list[Agent]:
@@ -28,24 +30,35 @@ class Topology:
             result.append(node)
         result.append(self.proxy)
         return result
-    
+
     def boot(self, verbose=True):
         """
-        Start up the topology
+        Start up every node in the topology
         """
-        if verbose:
-            print("Cleaning up...")
-        # Kill any patroni servers
-        os.system('pkill -9 -f "bin/patroni"')
+        if self.is_local:
+            if verbose:
+                print("Cleaning up local instances...")
+            # Kill any etcd/patroni servers
+            os.system('pkill -9 etcd')
+            os.system('pkill -9 -f "bin/patroni"')
 
         if verbose:
             print("Booting topology...")
         for agent in tqdm(self.agents, disable=not verbose):
             agent.boot(topology=self.config)
+    
+    def stop(self, verbose=True):
+        """
+        Stop every node in the topoloty
+        """
+        if verbose:
+            print("Stopping topology...")
+        for agent in tqdm(self.agents, disable=not verbose):
+            agent.stop()
 
 
 if __name__ == "__main__":
-    top = Topology("config/topology.local.yml", True)
-    encoded = jsonpickle.encode(top.config)
-    thawed = jsonpickle.decode(encoded)
+    top = Topology("config/topology.local.yml", is_local=True)
     top.boot()
+    time.sleep(2)
+    top.stop()
